requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.38.0
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.0.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
statsmodels==0.13.5
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
gunicornwaitress-serve --host 127.0.0.1 hello:app
fcntl
